[Bugfix] `add_*_stream` function errors hidden bug
The API methods for rs2_software_sensor_add_*_stream can return an
error in various cases. In the old implementation, the returned error code
was not checked before sending the result to the `stream_profile` constructor.
This was causing the originating error to get swallowed and made debugging
the actual problem harder. This fixes this method so that the
first error is reported and not subsequent errors.